requirement.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
Markdown==3.3.7
mysql-connector-python==8.0.29
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.1.0
